Added limit support to selectAll

+++ modules/cppql-typed/include/cppql-typed/expressions/expression_limit.h
@@ -11,8 +11,8 @@ namespace sql
 {
     struct LimitExpression
     {
-        size_t limit  = 0;
-        size_t offset = 0;
+        int64_t limit  = 0;
+        int64_t offset = 0;
 
         [[nodiscard]] std::string toString() const { return std::format("LIMIT {0} OFFSET {1}", limit, offset); }
     };
